Multi-threading: Improving interruption event handling. Removing InterruptionException from the Scraper public methods. When Thread.interrupt() occurs it sets the scrapper status to STATUS_STOPPED and breaks the execution. The thread interruption status is preserved after returning from Scraper.execute() so a user could optionally check it.
It's important to emphasize in the documentation the importance of the correct thread interruption status handling by the plug-in developers. Every time the interruption exception is cleared (either by catching InterruptedException or by invoking Thread.isInterrupted(true) one must call Thread.interrupt() again to restore the interruption status and not to hide it form the caller in any way!

git-svn-id: http://svn.code.sf.net/p/web-harvest/code/trunk@366 a57e4815-81e3-4922-91ae-0d9c2f5eb974